sustainability-api/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Evstifeev Roman", email = "someuniquename@gmail.com"},
]
dependencies = [
    "fastapi>=0.78.0",
    "ujson>=5.3.0",
    "mangum>=0.15.0",
    "django-environ>=0.9.0",
]
requires-python = ">=3.9"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.scripts]
_.env_file = "env-local"

[tool.pdm.dev-dependencies]
dev = [
    "uvicorn>=0.17.6",
    "awscli-local>=0.20",
    "awscli>=1.25.17",
    "ipdb>=0.13.9",
]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
